Add option to raise Exception on empty strings
Many JSON parsers (including Ruby's JSON module and Javascripts `JSON.parse`),
raise exceptions when only whitespace is given as an input string. This patch
adds an optional configuration parameter to raise an exception when an empty
string is given as an input.

This change also updates the Mimic configuration to use this by default, which
corresponds to the behavior of `JSON.parse`. This can be illustrated by running:

```
$ ruby -rjson -e "JSON.parse('   ')"
/home/ianks/.gem/ruby/2.3.1/gems/json-2.0.3/lib/json/common.rb:156:in `parse': 743: unexpected token at '' (JSON::ParserError)
	from /home/ianks/.gem/ruby/2.3.1/gems/json-2.0.3/lib/json/common.rb:156:in `parse'
	from -e:1:in `<main>'
```

Oj now raises a similar exception:

```
$ ruby -roj -e "Oj.mimic_JSON; JSON.parse('   ')"
-e:1:in `parse': unexpected character at line 1, column 3 [parse.c:606] (JSON::ParserError)
	from -e:1:in `<main>'
```