Yield next offset when parsing multiple objects
Pass the next offset as a second block argument when a block is passed
to `Oj.load`.

This allows an end-user to know what had been successfully parsed if an
error occurs after some objects have been parsed successfully, which is
useful if the stream being parsed may be incomplete (i.e. it's being
read from somewhere else that doesn't buffer on a per-object basis).