Fix stack limits with unlimited stacks (#522)
The maximum size of the stack of a process can be set to be unlimited;
in this case, the calculated stack limit in oj itself basically allows
no extra stack for nested JSONs, which are not parsed.

As solution, do not set a limit in case the resource limit result is
RLIM_INFINITY (i.e. not set).