Disable SSE 4.2 compile instructions by default (#806)
Compiling with `-msse4.2` is problematic because it may cause a number
of other SSE instructions to be used. For example:

```
$ ./elfx86exts oj.so
MODE64 (call)
CMOV (cmovne)
SSE2 (movq)
SSE41 (pinsrq)
SSE1 (movups)
SSSE3 (pshufb)
SSE3 (movddup)
SSE42 (pcmpestri)
CPU Generation: Unknown
```

The runtime check in #795 is therefore not sufficient to prevent SSE
instructions from leaking elsewhere in the shared library. As a
result, loading Oj on older machines with these libraries will result
in a "illegal instruction" crash.

Introduce a `--with-sse42` flag that can be used to configure this
gem.