Avoid using RARRAY_PTR
This macro marks the array as not protected by a write barrier
which has negative effects on GC performance.

As far as I could tell, all the uses in Oj are purely for read
so we can use `RARRAY_CONST_PTR` which is effectively the same
but doesn't unprotect the array.