Make invalid Unicode data raise when encoding through Oj::Rails::Encoder (#912)
* Actually run activesupport7 tests with Oj

These tests were not even loading Oj::Rails; they were definitely not
actually testing the Oj rails shim.

* Raise on invalid unicode in rails mode mimmicry

Activesupport & JSON gem will raise an exception when trying to an
encode an object containing a string with invalid byte sequences for the
string's encoding. Oj correctly raises if escaspe_html_entites_in_json
is enabled, but if that's disabled, the invalid byte sequence is copied
directly to the output.

Use the same logic to validate unicode in that case as well.