bench: Avoid allocating an extra hash when not necessary
`JSON.parse(@json, symbolize_names: @symbolize)` allocates an extra hash,
which isn't necessary when `symbolize` is `false`.

In the grand scheme of things, it's not a big deal, but on this sort
of micro-benchmarks that can easily accound for a huge part of the runtime
just because of the increased GC pressure.